requirements.txt
Flask==2.1.2
Werkzeug==2.1.2
spacy==3.2.3
PyMuPDF==1.18.17
python-docx==0.8.11
transformers==4.28.1
torch==1.11.0
